// src/entity/verb.tensens.enum.ts
export enum VerbTensesEnum {
    FutureQuestion = 'FutureQuestion',
    FutureStatement = 'FutureStatement',
    FutureNegative = 'FutureNegative',
    PresentQuestion = 'PresentQuestion',
    PresentStatement = 'PresentStatement',
    PresentNegative = 'PresentNegative',
    PastQuestion = 'PastQuestion',
    PastStatement = 'PastStatement',
    PastNegative = 'PastNegative'
}